refactor(renderer): own the renderer theme settings
